requirements.txt
setuptools==59.5.0

# local package
-e .

# external requirements
flake8
python-dotenv>=0.5.1
tqdm
pytest
hydra-core
omegaconf
fire

# project requirements
numpy
scipy
pandas
matplotlib
seaborn

pyro-ppl==1.4.0
laplace-torch
tensorboard
pytorch_lightning

-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.2+cu113
torchvision==0.11.3+cu113
torchaudio==0.10.2+cu113
